+++ packages/fetch/test/helpers/untyped-client.ts
@@ -8,26 +8,4 @@ export const client = untypedDredgeFetch().extends({
     text: "text/plain",
   },
   prefixUrl,
-
-  dataSerializers: {
-    "application/json": async ({ data }) => {
-      return JSON.stringify(data);
-    },
-    "text/plain": async ({ data }) => {
-      if (typeof data === "string") return data;
-      return "";
-    },
-  },
-  bodyParsers: {
-    "application/json": async ({ text }) => {
-      const payload = await text();
-      if (!payload) return;
-      return JSON.parse(payload);
-    },
-    "text/plain": async ({ text }) => {
-      const payload = await text();
-      if (!payload) return;
-      return payload;
-    },
-  },
 });