+++ packages/types/CHANGELOG.md
@@ -1,5 +1,14 @@
 # dredge-types
 
+## 0.4.0
+
+### Minor Changes
+
+- Fix type issues in client when There exist for static and dynamic path in same position. like /user/:id and /user/admin
+- static path no longer has duplicate path inference like this
+  - Before: `':/user/admin' | '/user/admin`
+  - After: `'/user/admin'`
+
 ## 0.3.0
 
 ### Minor Changes